Crate dioxus_document

Structs

Eval

Communicating with JavaScript

LinkProps

MetaProps

Props for the Meta component

NoOpDocument

A document that does nothing

ScriptProps

StyleProps

TitleProps

Enums

EvalError

Represents an error when evaluating JavaScript

ExtractSingleTextNodeError

An error that can occur when extracting a single text node from a component

Traits

Document

A provider for document-related functionality.

Evaluator

The platform’s evaluator.

Functions

Link

Render a link tag into the head of the page.

Meta

Render a meta tag into the head of the page.

Script

Render a script tag into the head of the page.

Style

Render a style or link tag into the head of the page.

Stylesheet

Render a link tag into the head of the page with the stylesheet rel. This is equivalent to the Link component with a slightly more ergonomic API.

Title

Render the title of the page. On web renderers, this will set the title in the head. On desktop, it will set the window title.

create_element_in_head

Create a new element in the head with javascript through the Document::eval method

document

Get the document provider for the current platform or a no-op provider if the platform doesn’t document functionality.

eval

Evaluate some javascript in the current document

Type Aliases

DocumentContext

A context for the document